Add files via upload
You will need Windows 8 or higher for the .exe to work.
If you are on Ubuntu/Debian, the pre-compiled Linux file included should work.
If you are on any other Linux or MacOS, download and run the index.py file from the src repo (You can download a .zip from the "Code" button) and run it with python3. You will also need to install some dependencies with pip, you can see them here.
Open application
If Windows tells you that it is not secure, it is because I can't be bothered to pay $200 for a digital license to verify my application. If you do not trust the .exe file, you can always run the Python source code instead, both do the same thing to be honest.
Paste your Discord token by copying it and right-clicking on the application to paste (or CTRL + V, some support it) and DO NOT share the token, please.
Note that to make it respond to your commands, you will need to keep the application/script window alive as it is what runs the bot to begin with. After you are done with everything, you can safely just close it.
When the bot has logged in, copy the invite link given to you by highlighting the URL and right-clicking, then go to a browser and paste the URL.
Invite the bot in to your server
Use the /hello command provided to you
NOTE: I recommend using an alt account to do this step, you do not have to, however it is best to do it, just to be safe.
Enter to https://discord.com/developers/active-developer and claim your badge.
It might take 24 hours before you are able to claim at all, so please take time to wait a bit if that is the case.